Add support for filtering by email (WIP)
This adds some new settings and code to allow administrators to filter
the allowable email addresses for oauth login. It doesn't work quite
yet, specifically the part that matches the oauth emails with the
existing database emails.

This will also basically only work with GitHub, because I don't think
that having a json list of the available emails for a user is a standard
oauth feature.